fix: swallow precondition/conflict on shared-ingress Cosmos writes; split controller tests
Address PR #6894 review feedback:

- create_shared_ingress_read_desire_controller: treat IsPreconditionFailedError on Replace as a no-op (Conflict on Create was already handled), matching the EnsureReadDesire pattern in backend/pkg/kubeapplierhelpers.
- shared_ingress_reporting_controller: treat IsPreconditionFailedError on the ManagementCluster Replace as a no-op; a newer generation re-triggers the sync.
- Split the combined controller_test.go into one file per controller (create_shared_ingress_read_desire_controller_test.go, shared_ingress_reporting_controller_test.go) with shared helpers in suite_test.go, and add precondition-failed-on-Replace cases for both controllers.
